refactor(server): logger - handler duplication
