+++ minecat/cogs/websocket/server.py
@@ -4,14 +4,13 @@
 
 from __future__ import annotations
 
-from logging import INFO, Formatter, LoggerAdapter, getLogger
-from logging.handlers import TimedRotatingFileHandler
 from typing import TYPE_CHECKING
 
 from botbase import CogBase
 from orjson import loads
 from websockets.server import serve
 
+from ._logger import logger
 from ._server import WebSocketServer
 
 if TYPE_CHECKING:
@@ -19,31 +18,6 @@ if TYPE_CHECKING:
 
     from minecat.__main__ import Minecat
 
-
-class UUIDAdapter(LoggerAdapter):
-    def process(self, msg, kwargs):
-        try:
-            websocket = kwargs["extra"]["websocket"]
-        except KeyError:
-            return msg, kwargs
-
-        xff = websocket.request_headers.get("CF-Connecting-IP")
-        uuid = getattr(websocket, "uuid", "unknown".ljust(22))
-        return f"{uuid} {xff}: {msg}", kwargs
-
-
-raw_logger = getLogger("minecat.cogs.websocket")
-raw_logger.setLevel(INFO)
-h = TimedRotatingFileHandler("./logs/ws/io.log", when="midnight")
-h.setFormatter(
-    Formatter(
-        "%(levelname)-7s %(asctime)s %(filename)12s:%(funcName)-28s: %(message)s",
-        datefmt="%H:%M:%S %d/%m/%Y",
-    )
-)
-h.namer = lambda name: name.replace(".log", "") + ".log"
-raw_logger.addHandler(h)
-logger = UUIDAdapter(raw_logger)
 
 
 class Server(CogBase["Minecat"]):
